% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.eoa.r
\name{model.matrix.eoar}
\alias{model.matrix.eoar}
\title{model.matrix.eoa}
\usage{
\method{model.matrix}{eoar}(object, data, ...)
}
\arguments{
\item{object}{An object of class \code{eoar}. In particular,
this object must have a $terms component.}

\item{\dots}{Included for compatability with other methods}
}
\value{
The design matrix for an \code{eoar} object with all the factor
levels and interactions expanded.
}
\description{
Model matrix method for eoar objects produced
by the \code{eoar} function.
}
\author{
Trent McDonald
}
