% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eoa.r
\name{plot.eoar}
\alias{plot.eoar}
\title{plot.eoar - Plot rate estimates from an \code{eoar} model}
\usage{
\method{plot}{eoar}(obj, xvar = NULL, main = NULL)
}
\arguments{
\item{obj}{An object of class "eoar".  Usually output by the
routine \code{eoar}.}

\item{xvar}{The x-axis variable to plot.  Points in the plot
are carcasses/g/numTurbines ~ xvar and line in plot is lambda ~ xvar.
xvar must be in the model you want and in the data frame contained
in the model object. All other covariates in the model are held
constant at their mean value.}

\item{main}{The main title of the plot}
}
\description{
Plot method for \code{eoar} objects.
}
\seealso{
\code{\link{plot.Mest}}
}
\author{
Trent McDonald
}
