% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostPredictEoAR.R
\name{PostPredictEoAR}
\alias{PostPredictEoAR}
\title{Make posterior predictions}
\usage{
PostPredictEoAR(object, newdata, newoffset, beta.params = NULL, nyears = 1)
}
\arguments{
\item{object}{A fitted \code{eoar} object.}

\item{newdata}{A data frame of covariates for new observations.}

\item{newoffset}{A vector offsets for new observations.}

\item{beta.params}{A data frame containing, at a minimum, two columns 
named \code{$alpha} and \code{$beta}. These are the alpha and beta 
parameters of a Beta distribution that determine the g-values in 
each "cell". Length of \code{$alpha} and \code{$beta} is either one, 
in which case g is assumed constant over "cells", or one element 
per "cell".}

\item{nyears}{A scalar for the number of years to make predictions for.}
}
\value{
A list of posterior predictive distributions.
}
\description{
Make posterior predictions
}
\author{
Andrew Tredennick
}
