% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.eoa.r
\name{print.eoa}
\alias{print.eoa}
\alias{print.eoar}
\title{print.eoar - Print a Evidence of Absence Regression model}
\usage{
\method{print}{eoar}(x, ...)
}
\arguments{
\item{x}{An estimated eoar object from \code{eoar}.}

\item{\dots}{Included for compatibility with other print methods.  Ignored here.}
}
\value{
The input value of \code{obj} is invisibly returned.
}
\description{
Print method for EoAR models produced by \code{eoar},
which are of class \code{eoar}.
}
\examples{

}
\seealso{
\code{\link{summary.eoar}}
}
\author{
Trent McDonald, WEST Inc. \email{tmcdonald@west-inc.com}
}
\keyword{models}
